+++ packages/sit-onyx/src/components/OnyxPageLayout/examples/SidebarExample.vue
@@ -1,28 +1,31 @@
 <script setup lang="ts">
-import { OnyxHeadline, OnyxPageLayout } from "../../..";
+import { OnyxAppLayout, OnyxHeadline, OnyxLink, OnyxPageLayout, OnyxSidebar } from "../../..";
 </script>
 
 <template>
-  <OnyxPageLayout class="page">
-    <template #sidebar>
-      <aside class="sidebar">OnyxSidebar component coming soon! Stay tuned.</aside>
-    </template>
+  <OnyxAppLayout>
+    <OnyxPageLayout>
+      <template #sidebar>
+        <OnyxSidebar label="Example sidebar">
+          <div class="sidebar__content">
+            For sidebar examples, see:
+            <OnyxLink href="/?path=/docs/navigation-sidebar--docs" target="_parent">
+              OnyxSidebar
+            </OnyxLink>
+          </div>
+        </OnyxSidebar>
+      </template>
 
-    <!-- page content -->
-    <OnyxHeadline is="h1">Page content</OnyxHeadline>
-  </OnyxPageLayout>
+      <!-- page content -->
+      <OnyxHeadline is="h1">Page content</OnyxHeadline>
+    </OnyxPageLayout>
+  </OnyxAppLayout>
 </template>
 
 <style lang="scss" scoped>
-.page {
-  // only needed for this example. Can be removed if used together with the OnyxAppLayout
-  height: 100vh;
-}
-
 .sidebar {
-  border-right: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
-  background-color: var(--onyx-color-base-background-blank);
-  height: 100%;
-  padding: var(--onyx-density-md);
+  &__content {
+    padding: var(--onyx-sidebar-padding);
+  }
 }
 </style>