+++ packages/sit-onyx/src/components/OnyxPageLayout/examples/SidebarWithPageFooterExample.vue
@@ -1,27 +1,37 @@
 <script setup lang="ts">
-import { OnyxHeadline, OnyxPageLayout } from "../../..";
+import { OnyxAppLayout, OnyxHeadline, OnyxLink, OnyxPageLayout, OnyxSidebar } from "../../..";
 </script>
 
 <template>
-  <OnyxPageLayout class="page" footer-alignment="page">
-    <template #sidebar>
-      <aside class="sidebar">OnyxSidebar component coming soon! Stay tuned.</aside>
-    </template>
+  <OnyxAppLayout>
+    <OnyxPageLayout footer-alignment="page">
+      <template #sidebar>
+        <OnyxSidebar label="Example sidebar">
+          <div class="sidebar__content">
+            For sidebar examples, see:
+            <OnyxLink href="/?path=/docs/navigation-sidebar--docs" target="_parent">
+              OnyxSidebar
+            </OnyxLink>
+          </div>
+        </OnyxSidebar>
+      </template>
 
-    <!-- page content -->
-    <OnyxHeadline is="h1">Page content</OnyxHeadline>
+      <!-- page content -->
+      <OnyxHeadline is="h1">Page content</OnyxHeadline>
 
-    <template #footer>
-      <!-- we are using the "onyx-grid-container" class here to sync the footer with the global application max-width, alignment and horizontal padding -->
-      <footer class="footer onyx-grid-container">Footer content</footer>
-    </template>
-  </OnyxPageLayout>
+      <template #footer>
+        <!-- we are using the "onyx-grid-container" class here to sync the footer with the global application max-width, alignment and horizontal padding -->
+        <footer class="footer onyx-grid-container">Footer content</footer>
+      </template>
+    </OnyxPageLayout>
+  </OnyxAppLayout>
 </template>
 
 <style lang="scss" scoped>
-.page {
-  // only needed for this example. Can be removed if used together with the OnyxAppLayout
-  height: 100vh;
+.sidebar {
+  &__content {
+    padding: var(--onyx-sidebar-padding);
+  }
 }
 
 .footer {
@@ -29,12 +39,5 @@ import { OnyxHeadline, OnyxPageLayout } from "../../..";
   background-color: var(--onyx-color-base-background-blank);
   padding-top: var(--onyx-density-md);
   padding-bottom: var(--onyx-density-md);
-}
-
-.sidebar {
-  border-right: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
-  background-color: var(--onyx-color-base-background-blank);
-  height: 100%;
-  padding: var(--onyx-density-md);
 }
 </style>