+++ packages/sit-onyx/src/components/OnyxSidebar/examples/DefaultExample.vue
@@ -1,0 +1,33 @@
+<script setup lang="ts">
+import { OnyxAppLayout, OnyxButton, OnyxHeadline, OnyxPageLayout, OnyxSidebar } from "../../..";
+</script>
+
+<template>
+  <OnyxAppLayout>
+    <OnyxPageLayout>
+      <template #sidebar>
+        <OnyxSidebar label="Example sidebar">
+          <template #header> Header content </template>
+
+          <div class="sidebar__content">Body content</div>
+
+          <template #footer>
+            <OnyxButton color="neutral" label="Button" />
+            <OnyxButton label="Button" />
+          </template>
+        </OnyxSidebar>
+      </template>
+
+      <!-- page content -->
+      <OnyxHeadline is="h1">Page content</OnyxHeadline>
+    </OnyxPageLayout>
+  </OnyxAppLayout>
+</template>
+
+<style lang="scss" scoped>
+.sidebar {
+  &__content {
+    padding: var(--onyx-sidebar-padding);
+  }
+}
+</style>