+++ packages/nuxt-docs/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sit-onyx/nuxt-docs
 
+## 1.0.0-beta.26
+
+### Patch Changes
+
+- Updated dependencies [3288513]
+  - sit-onyx@1.0.0-beta.213
+  - @sit-onyx/nuxt@1.0.0-beta.215
+
+## 1.0.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [fe38f20]
+  - sit-onyx@1.0.0-beta.212
+  - @sit-onyx/nuxt@1.0.0-beta.214
+
+## 1.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [e2f1083]
+  - sit-onyx@1.0.0-beta.211
+  - @sit-onyx/nuxt@1.0.0-beta.213
+
+## 1.0.0-beta.23
+
+### Patch Changes
+
+- Updated dependencies [a39d444]
+  - sit-onyx@1.0.0-beta.210
+  - @sit-onyx/nuxt@1.0.0-beta.212
+
 ## 1.0.0-beta.22
 
 ### Patch Changes